♻️ DNT18Z release: split apply preflight helpers
